sip/execution_control/processing_controller/processing_block_controller/requirements.txt
docker==3.3.0
jinja2==2.10
PyYaml==3.13
redis==2.10.6
celery==4.2.1
skasip-config-db==1.1.5
skasip-docker-swarm==1.0.5
skasip-logging==1.0.14
